requirements.txt
streamlit
langchain
langchain-community
langchain-mistralai
python-dotenv
faiss-cpu
pymupdf
